# Node16 add: implement SIMD search for insert position